Restricted Page Commit

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -15,7 +17,10 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
-    
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method =='POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -24,6 +29,7 @@ def loginPage(request):
             user = User.objects.get(username=username)
         except:
             messages.error(request, 'User does not exist')
+            
 
         user = authenticate(request, username=username, password=password)
 
@@ -62,6 +68,7 @@ def room(request, pk):
  
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -74,9 +81,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -87,8 +99,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
